usb-key: add pictures to prove a point
That's way much effort I guess.
I should've been instead focusing on the means on removing
the annoying voice prompts instead.
Though this is actually important to know so that you could
flawlessly enter the USB bootloader to be able to dump
the firmware to begin with, and then flash the modified
firmware back. So that's not a waste of effort!